feat: extract generic funtions
