+++ lib/analyzer/applications/python/index.ts
@@ -1,9 +1,4 @@
-import { getPythonApplicationFiles } from "./common";
 import { pipFilesToScannedProjects } from "./pip";
 import { poetryFilesToScannedProjects } from "./poetry";
 
-export {
-  getPythonApplicationFiles,
-  pipFilesToScannedProjects,
-  poetryFilesToScannedProjects,
-};
+export { pipFilesToScannedProjects, poetryFilesToScannedProjects };
